Add files via upload
The primary objective of this analysis was to predict the content rating (Y) of Netflix shows based on various features (X) such as the director, genre, duration, and release year of the shows. Content rating prediction would help in classifying shows into appropriate age and maturity groups, ensuring viewers are presented with content that is suitable for their age group.