Add files via upload
This dataset offers an in-depth examination of salary trends among engineering graduates, covering a wide range of variables including the
engineering specialty, alma mater, graduation year, job location, employer, position held, initial and current salaries, and industry tenure. The
dataset's purpose is to facilitate the analysis of salary patterns and trends for those with engineering degrees.